Created C echo client, minor changes in server code.
I created a simple C client that creates a socket which connects to
the server. The echo client then gets user input and writes to the
server with send() call. It receives the server response call with
recv(). It all works ok, no major crazyness. Should probably
shape up the C code, align it with GNU or linux kernel standard.